- add authorization for deleting pet   - ensure that pet being deleted is associated with current user

+++ src/actions/actions.ts
@@ -94,6 +94,11 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 }
 
 export async function deletePet(petId: unknown) {
+  // authentication check
+  const session = await auth();
+  if (!session?.user) redirect("/login");
+
+  // validation
   const validatedPetId = petIdSchema.safeParse(petId);
   if (!validatedPetId.success) {
     return {
@@ -101,6 +106,17 @@ export async function deletePet(petId: unknown) {
     };
   }
 
+  // authorization check
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    },
+  });
+  if (!pet) return { message: "Pet not found..." };
+  if (pet.userId !== session.user.id)
+    return { message: "Current user not authorized..." };
+
+  // database mutation
   try {
     await prisma.pet.delete({
       where: {
